fn-http/requirements.txt
parliament-functions==0.1.0
requests
jsonify
flask